# db/go/sales_management/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "sqlx + Gin による販売管理システムの REST API",
        "title": "販売管理システム API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/products": {
            "get": {
                "description": "すべての商品を取得します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "商品一覧取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schema.ProductResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "新しい商品を作成します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "商品作成",
                "parameters": [
                    {
                        "description": "商品作成リクエスト",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/schema.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/products/{prodCode}": {
            "get": {
                "description": "指定された商品コードの商品を取得します",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "商品詳細取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品コード",
                        "name": "prodCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.ProductResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "指定された商品コードの商品を更新します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "商品更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品コード",
                        "name": "prodCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "商品更新リクエスト",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.UpdateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "指定された商品コードの商品を削除します",
                "tags": [
                    "products"
                ],
                "summary": "商品削除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品コード",
                        "name": "prodCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "schema.CreateProductRequest": {
            "type": "object",
            "required": [
                "fullName",
                "name",
                "poPrice",
                "prodCode",
                "supCode",
                "unitPrice"
            ],
            "properties": {
                "categoryCode": {
                    "type": "string",
                    "example": "CAT001"
                },
                "fullName": {
                    "type": "string",
                    "example": "黒毛和牛サーロインステーキ 200g"
                },
                "kana": {
                    "type": "string",
                    "example": "クロゲワギュウサーロイン"
                },
                "name": {
                    "type": "string",
                    "example": "サーロイン"
                },
                "poPrice": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 3500
                },
                "prodCode": {
                    "type": "string",
                    "example": "PROD00001"
                },
                "supCode": {
                    "type": "string",
                    "example": "COMP0011"
                },
                "unitPrice": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 5000
                }
            }
        },
        "schema.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string",
                    "example": "詳細情報"
                },
                "error": {
                    "type": "string",
                    "example": "エラーメッセージ"
                }
            }
        },
        "schema.ProductResponse": {
            "type": "object",
            "properties": {
                "categoryCode": {
                    "type": "string",
                    "example": "CAT001"
                },
                "fullName": {
                    "type": "string",
                    "example": "黒毛和牛サーロインステーキ 200g"
                },
                "kana": {
                    "type": "string",
                    "example": "クロゲワギュウサーロイン"
                },
                "name": {
                    "type": "string",
                    "example": "サーロイン"
                },
                "poPrice": {
                    "type": "integer",
                    "example": 3500
                },
                "prodCode": {
                    "type": "string",
                    "example": "PROD00001"
                },
                "supCode": {
                    "type": "string",
                    "example": "COMP0011"
                },
                "unitPrice": {
                    "type": "integer",
                    "example": 5000
                }
            }
        },
        "schema.UpdateProductRequest": {
            "type": "object",
            "properties": {
                "categoryCode": {
                    "type": "string",
                    "example": "CAT001"
                },
                "fullName": {
                    "type": "string",
                    "example": "黒毛和牛サーロインステーキ 200g"
                },
                "kana": {
                    "type": "string",
                    "example": "クロゲワギュウサーロイン"
                },
                "name": {
                    "type": "string",
                    "example": "サーロイン"
                },
                "poPrice": {
                    "type": "integer",
                    "example": 3500
                },
                "supCode": {
                    "type": "string",
                    "example": "COMP0011"
                },
                "unitPrice": {
                    "type": "integer",
                    "example": 5000
                }
            }
        }
    }
}